Adapt to respect new Firefox API
Firefox seems to set window.chrome to window.browser so it is unreliable to
check for winow.chrome to be set and assume it to be a Chrome browser.
Chrome however does not set window.browser for now so we can use that
to identify whether the current browser is indeed Chrome.
Firefox also made changes to the devtools.inspectedWindow API which is
more in line with Chrome now.
See https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/devtools/inspectedWindow/eval#syntax
for further information.